+++ CM.Text/BusinessMessaging/Model/MultiChannel/IRichMessage.cs
@@ -1,0 +1,12 @@
+﻿using JetBrains.Annotations;
+
+namespace CM.Text.BusinessMessaging.Model.MultiChannel
+{
+    /// <summary>
+    ///     One element in a <see cref="RichContent.Conversation" />
+    /// </summary>
+    [PublicAPI]
+    public interface IRichMessage
+    {
+    }
+}