+++ CM.Text/Common/Constant.cs
@@ -2,6 +2,8 @@
 {
     internal static class Constant
     {
-        internal static readonly string TextSdkReference = "text-sdk-dotnet-" + typeof(TextClient).Assembly.GetName().Version;
+        internal static readonly string TextSdkReference = "text-sdk-dotnet-" +
+                                                           typeof(TextClient).Assembly.GetName()
+                                                               .Version;
     }
 }